Make admin pages more robust when data missing
We currently get errors when data that _should_ be there isn't.
Sometimes this is because of bad data and other times it's just a little
fragile.

Moving the assignment of instance variables to the controllers means we
aren't traversing them in the template and makes everything a little
easier to follow.