CPDLP-1924 Add method which calculates the correct EHCO schedule
For each cohort, we need to return the correct EHCO schedule depending on the time the NPQ application is accepted.
Calculate the correct schedule depending on the time range. If the schedule does not exist or is outside the time
range raise on error.

Current data in production is not the same for each cohort. For cohorts 2022 onwards we have all the schedules present. For 2021 we only have one. Default to that schedule if the cohort passed in is less than 2022