Add some extra data to the default seeds
In the legacy seeds we automatically created some logins that were
sometimes used for testing in dev. I didn't realise people were actually
using them but they are (occasionally!)

This change adds a few back and introduces the
NewSeeds::Scenarios::InductionCoordinator scenario. The requirements
requested that the following users are present:

* cpd-test+school-1@digital.education.gov.uk - School primary contact (cannot sign in)
* cpd-test+school-2@digital.education.gov.uk - School primary contact (cannot sign in)
* cpd-test+school-3@digital.education.gov.uk - School primary contact (cannot sign in, with expired token)
* cpd-test+tutor-1@digital.education.gov.uk - School induction tutor (core induction programme)
* cpd-test+tutor-2@digital.education.gov.uk - School induction tutor (full induction programme)
* lead-provider@example.com - Lead Provider
* school-leader@example.com - School Leader