Fix NPQ amend cohort flakey spec
We amended the schedule defaults recently, meaning that the schedule
that might be selected in the test can be incorrect. Use the same
logic to select a schedule by "schedule_for" to ensure it matches the
returned schedule in the factory.

Also create the schedules in the previous year to ensure they are
available as well.