Build induction coordinators when seeding data
Previously schools were built without an induction coordinator. So we
can test various journies more easily the following changes have been
made:

* make user emails end in @example.com by default (allowing us to log in
  with any account in dev)
* ensure we generate induction coordinator profiles for schools when
  seeding

The process for testing out a school journey as an IC would be to log in
with admin@example.com, choose a school, copy the email address and
either log out and back in with it or to log in in a private browser
window.