Add NPQ API for user creation that supports GAI
This commit adds in a new API endpoint at  POST api/v1/npq/users. This API endpoint is a replacement for POST api/v1/users which also is also in place for creating users.

The difference between the two is that in the existing API the accepted params are exclusively email and full_name whereas in the new API it also accepts get_an_identity_id. This seems like it would be backwards compatible however it is not due to certain changes in behaviour such as:
- Which errors and status codes are returned, a move to 200s+201s and differing 400s.
- A move to where “recreating” a user no longer returns an error and instead just returns the same user

I also want to be careful with the change, leaving the old API in place and adding a new namespaced one ensures backwards compatibility is preserved and that the usage of the new API can be rolled back without needing to rollback multiple systems.

# Conflicts:
#	db/schema.rb