Add after build step to allow local preview
When running middleman locally the hardcoded paths for JavaScript and
Stylesheets make previewing changes difficult.

Now, instead of hardcoding a api-reference prefix we'll just use
javascripts and stylesheet directories in source and when building
(using `middleman build`) run an additional step which copies the
directories to api-reference, so the site will render correctly when
served from within Rails too.