Add api/v1/npq/previous_funding#show
This API is built as a replacement for api/v1/npq-funding/:trn.

The primary goal of building this replacement is to introduce an API that doesn’t depend on the existence of a TRN for it to function.

The existing API embedded the TRN being queried against into the URL itself. The new API doesn’t include any identifiers in the URL, instead receiving the user identifier data in the same way as course identifier information via query parameters.

This allows us to break out from a model where you have to have a TRN to where you can use multiple different kind of identifiers.

Old API formatting example:
/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership

New API formatting example:
/api/v1/npq/previous-funding?npq_course_identifier=npq-senior-leadership&trn=1234567