Stop returning Sentry object when calculating qualification type
When we can't find an NPQ Outcome qualification type, we are logging
this to Sentry. Since this is the final step in the method, the return
value from the capture method is set to the qualification type and
is attempted to be sent to the qualification API.

This returns an error and logs a large request body. Instead send
an empty qualification type to log an the correct error from the API
instead