Add new column for uploaded partnership URNs
Instead of using S3 to store the provided CSV which only contains a list
of URNs we can simply save them directly in the database. This is
considerably simpler and will make the migration to Azure easier as we
won't need to transfer any files or do any kind of manual sync.