Check induction record schedule before raising validation error
There are some cases where the profile schedule will be the same as
the new schedule we are attempting to change to, however the induction
record has a different schedule.

This can happen in transfers, when a profile is amended after a participant
has transfered, leaving the induction record with an incorrect schedule,
but the profile with the correct one.

Allow the change to happen to align all schedules without raising a
validation error