Avoid n+1 queries when checking school partnerships in a cohort
The current implementation uses a query which causes n+1s. Instead
calculate the valid partnerships in memory to avoid the extra query.

The risks are low as usually schools have 1/2 partnerships, and a max
of 6 from what I could find. This method is used in one other tool
internally so shouldn't have much impact.