Add transfers serialiser
Add new transfers serialiser and build transfers service. The transfer
service traverses the leaving induction records, and then picks the next
joining induction inline, otherwise the transfer is an incomplete transfer.

After that is done for the first leaving record, we move to the next pair,
and reset the values.

Complexity is high but I couldn't find a better way to do this

This is used in the serialiser which builds transfers per user, by mapping
over all participant profiles.