Add an example delivery partner and school
This record should allow delivery partners to get a feel for how the
school interface works. Using it in the dev environment will ensure the
data is reseeded frequently so we can tweak and adjust what's seeded
without needing to worry about manually applying changes in a more
static env.