Split out summary_list_rows for finer control
When using the single content 'slot' we weren't able to place the
additional summary list rows within the summary list; or we placed extra
non-summary-list content inside it. Both are wrong.

Now we have an extra slot called summary_list_rows which can be used to
insert rows in the correct place.