Refresh the training record states every 2 mins
Refreshing concurrently takes about 40 seconds so a 2 minute window
should be sufficient. Refreshing a materialized view creates an exlusive
lock on the view, so multiple can't be run at the same time. The slight
concern is what happens if we slow down and many queue up (will they all
run or will it combine the multiple calls into one?)

We'll need to keep an eye on this and adjust the schedule if needed