Make determine_training_record_state perform query
After trying the materialized view approach we ran into some problems
where the refresh took too long when the server was under load. This
caused the refresh jobs to back up, putting the server under even more
load and eventually causing timeouts, delays and crashes.

Instead, this change implements the query in raw SQL. This has the
disadvantage of being ugly and requiring that we construct the query
manually with interpolation, but we can speed it up by targetting the
CTEs.

Now, DetermineTrainingRecordState.call returns a
DetermineTrainingRecordState::Record struct that matches the row format
for a TrainingRecordState.