Allow filtering of applications by participant_id
The API v3 schema surfaces the ability to filter NPQ applications by
`participant_id`, however we have not yet implemented the functionality.

Allow filtering on `participant_id` for v3 of the NPQ applications API.

Migrate to a query model for consistency with other v3 endpoints.

Request test added for `participant_id` filtering, but we may want to
clean up these controller tests at a later date so we're not over-testing
the query models. It looks like there's also scope to DRY up the query
models with a concern similar to `ApiFilter` and/or additional scopes on
the core models.