Add mentor_user_id to ParticipantDeclaration
In order to optimise and simplify queries that reference a
declaration `mentor_id` we are going to store the value on the
model directly. We will be populating this when the
declaration is created and backfill existing declarations.

In the `ParticipantDeclarationQuery` we compute the
`mentor_user_id` as an optimisation; to avoid a clash with te
new attribute name (as both `ParticipantDeclaration` and query
results can filter through the serializer) the computed
attribute has been renamed `transient_mentor_user_id` and will
be removed in a future commit.