Fix Induction::ChangeProgramme to remove mentor from the correct school
The problem is that the current_induction_record becomes the enroled
induction record, meaning the Mentors::ChangeSchool will add the mentor
to the new school and then remove them from the same school instead of
removing them from the old school.

Keeping the current induction record of the old school and re-using it,
fixes the issue.