Clarify induction record date edge cases
We have induction records in the database that are in an
invalid/unexpected state. Until we are able to clean up the
data and ensure consistency we need to make sure the latest
induction record logic is exepcted/documented.

Update `InductionRecord` scopes to prioritise records based
on the latest `end_date` if they have the same `start_date`.

Add test coverage to the `FindBy` class to cover the
following edge cases:

- Induction records with the same start date where one has a `nil` end date
- Induction records with the same start date where one has a later end date
- The expected scenario of the latest induction record having a `nil` end date