Optimise participant declarations query
I looked into a number of ways to optimise this query; the main bottleneck is
the Postgres sort operation in the query plan, which uses a whopping 46mb of
space and results in a disk write as our working memory is only ~4mb.

The high sort memory is due to a `UNION` over a lot of columns. We can reduce
the sort overhead by performing the `UNION` over the `id` columns and then
re-select the records from the result. This was more efficient, but less
efficient than performing a `UNION` over a subset of columns (that are used in
subsequent filtering) and using the resulting records directly instead of
re-selecting them. Our `ActiveRecordUnion` model was doing both; `UNION` over
all columns _and_ re-selecting the results for some reason. I opted to swap out
our `ActiveRecordUnion` service as it wasn't tested and there is a gem that
performs an optimal `UNION` available. The sort now uses ~12mb of space.

After the sort operation the next slowest operation is a nested loop due to the
amount of join statements. To avoid this I've optimised the joins (hopefully
without breaking anything) and only applied them when they're actually needed
(such as when filtering by cohort). We can't make up much performance-wise here
though.

The final optimisation was to remove the unnecessary `distinct` at the end of
the query; we do a `UNION` which internally performs distinct so this isn't
necessary. I tried `UNION ALL` (which doesn't do a distinct) and keeping a
manual call to `distinct` at the end but this was less efficient.

All of this results in almost ~70% performance improvement; from ~2.3s to
~700ms.