Optimize NPQApplicationsQuery
The query to retrieve `NPQApplications` is slow for two reasons:

1. When calculating `previously_funded?` for each record we retrieve all
previously funded applications, exclude the current application and then check
if there are any.
2. For each record we need to query the previously funded applications.

Fixing #1 is straight-forward; we can bring the exclusion logic into the query
and call `exists?` so we're only returning a boolean and not all the previous
records.

Fixing #2 is not so straightforward. To speed this query up we need to find the
previously funded applications as part of the query to retrieve all
applications. Most of the sub-query to do this is straight-forward; the only
difficulty is filtering by 'rebranded alternative' courses. Each course has a
set of these, as determined by `NPQCourse#rebranded_alternative_courses`. To be
able to dynamically filter on these across all records we first build a JSON
map of course ID -> alternative course IDs. We can then use this in a subquery
to check each application and determine if it has been previously funded. The
result is stored on a `transient_previously_funded` attribute of the model,
which can then be used to optimize the `previously_funded?` method of the
application during serialization.

These optimisations reduce the endpoint response time from ~3.8s to ~650ms.