Change "To be confirmed" to be more specific
Some users got in touch with support asking how they could confirm lead provider and delivery partner.

The text "To be confirmed"  did not make clear that this had to be confirmed by the lead provider. We’ve changed this to "Your lead provider needs to confirm this with us".