Put ZenDesk widget behind a feature flag
Usually we have no need for a feature flag as our support works by
email. However, on a couple of occasions we've run into problems where
our email address has stopped working (or the forwarding from O365 to
ZenDesk has broken) and that leaves us unable to support our users.

During the last major incident we added a ZenDesk widget that appears on
the page when a user is logged in. This was reverted after the incident
was closed, and then re-reverted to add it back during the second
incident.

Now, instead of re-re-reverting we'll hide it behind a flag, so it can
easily be enabled using the Rails console should the need arise.