Add a smoke test for review apps
There is a new smoke-test action which runs a new test script,
`bin/smoke`, that ensures an environment is working correctly.

We want to avoid doing anything too heavy in this step so aren't using
RSpec and driving a browser to hit multiple endpoints; instead we are
relying on /healthcheck to tell us if our application is
working. If we feel like we need other checks too, like to ensure
JavaScript or CSS have been built, we can do that in the
HealthcheckController too.

The `bin/smoke` script checks the following things:

* the endpoint is reachable and responds without error
* the app is deployed with the correct sha
* the database migrations have all been run