Export correct delivery partner in assurance report
The assurance report query currently takes the delivery partner from the
latest induction record. This may not be correct in all scenarios and is
inconsistent with how we select the delivery partner in other areas of
the application.

Instead, we want to use the delivery partner associated with the
declaration and only fallback to the latest induction record if it is not
set (which it may not be on some historical records).

General clean-up to the query specs.