Update unfunded mentor information
Replaced: 

Providers can view the names and email addresses of ‘unfunded mentors’ assigned to ECTs, and give these mentors appropriate access to learning platforms.

With:

A participant will have the same participant_id throughout the API. 

If a participant is a mentor and linked to an ECT, then the mentor_id that appears on the response for their ECT will be the mentor’s participant_id. 

If the mentor is an unfunded mentor, then the identifier will be their participant_id.

And...

Added: 

Using the unfunded mentors endpoints, 

At the start of the following par: 

Providers can view the names and email addresses of ‘unfunded mentors’ assigned to ECTs, and give these mentors appropriate access to learning platforms.