Minor updates to deduplication logic
Go to the participant detail view after deduplication (to more easily see the
result of the deduplication).

Ensure all induction records have an `end_date` - if the school transfer logic
isn't applied the other induction records were just being moved; we need to
ensure the latest gets an `end_date` in this scenario as well.

If an induction record on the duplicate already has an `end_date` we should
retain it rather than overwrite it.

Change the order of the updates so that we reconcile induction records last (as
the declaration reconciliation will attempt to read the induction records on
the duplicate).