Prevent primary/duplicate ids swapping in path on dedupe
The dedupe path was configured to be the opposite of the compare
path, which meant the ids were getting swapped in the URL on the
dry-run/dedupe. Whilst functionally this is fine and was working
it makes it confusing as I often copy/paste paths when I'm working
through the duplicates and it meant I was manually swapping them
by accident.

Update dedupe paths to use the same duplicate/primary profile ID
ordering as the compare path.