Update deduplicator to handle additional scenarios
We can now handle duplicates with a different lead provider but
the same school - providing there are no declarations on the
duplicate - by destroying the duplicate profile _without_
reconciling it with the primary.

After working through a change in training programme with Nathan
on a call we've decided we can probably handle these but they need
a warning to ensure we double check the ordering and scrutinise
them more closely.