Use API Token owner to identity `whodunnit` in change log versions
Errors have been reported when users tried to use the new feature Changelog.
The error reported that users have clicking on the changelog button, the user
was redirected to `page not found`

After connecting to the Box, I have noticed that the attribute `whodunnit` for
the first version of an NPQ Application is `npq_registration_application`, so
when we tried to find the user it blows up.