Change transfer direction in mentor to sit transfer
This commit changes the direction of the transfer operation between mentor and sit user. Previously, the mentor's identity was transferred to the sit user and the mentor user was destroyed. This has now been reversed. The sit user's identity is now transferred to the mentor user, and the sit user is destroyed.