Add a database populated check to the healthcheck
The healthcheck is used by the smoke tests to ensure that the
environments are in good shape before continuing the continuous
deployment process.

This check provides some extra safety by ensuring if we make changes to
resource names in terraform that result in them being deleted (#3992, #4008)
and recreated, the emptiness of the tables will cause a failure.