Add primary key to HABTM table
We have a `has_and_belongs_to_many` table
`cohorts_lead_providers`. By default Rails doesn't add a
primary key `id` column to this table (as it as a composite of
the foreign keys). The `dfe-analytics` gem, however, requires
an `id` column that is a primary key on all tables it sends to
BigQuery.