"Add check for privacy policy acceptance in transfer controller
This commit introduces a conditional check for privacy policy acceptance when a user is signed in after a transfer. This change is necessary to ensure that the acceptance of the privacy policy is only forced when it is required for the current user. Previously, the privacy policy acceptance was forced every time a user was signed in after a transfer, whether required or not."