Convert DetermineTrainingRecordState to Ruby
The `DetermineTrainingRecordState` was udpated to be a SQL query so that we
could use materialized views (for performance). We found that this locked up
our production database, so we ended up moving away from using the materialized
views.

The SQL is quite hard to follow and contains a lot of business logic. We would
like to move back to a Ruby implementation for ease of understanding and
potentially look at other ways to improve performance (such as supporting
multiple profiles or caching the training states in their own model/table).

Revert implementation to Ruby and update it to be consistent with the latest
SQL logic.