Support multiple participant profiles
We want to update the `DetermineTrainingRecordState` service so that it can
support calculating the training record states for multiple participants in a
more optimal way (as we are finding pages that call this service many times are
slow to load).

The approach taken is to abstract the query from the service itself and instead
take the query used in the controller which already contains most of the
filtering/logic we need. The query is then updated with some record-state
specific joins/subequeries that will further improve performance.

To begin with only the `AppropriateBodyStatusTag` view component has been
updated to accept the new `TrainingRecordState`. Going forward the plan is to
migrate the other view components and update the queries appropriately.

The current service has been renamed to `DetermineTrainingRecordStateLegacy`.