Move logic for reactivating profiles into dedicated service
The logic for how to reactivate was starting to impact the logic fo0r creating ne w records. This wasn’t a good move as it corrupted the core intent of the ECT/Mentors Create services into a mix of two core problems. Separating them out made things clearer.

Also, the code now will update emails and uses TRNs for profile recognition so that it doesn’t create duplicate profiles if you use a different email to the one the school previously used.