Change XL headings to L
Following a change to the GOV.UK Design System guidance, this drops all the h1 headings from Extra Large (XL) size, down to Large (L).

Where the heading is on a content page with subheadings, the subheadings are now set to Medium (M) for the h2 and small (S) for any h3s.