Add ChangeServiceFees oneoff service
Occasionally we need to adjust the servce fees for statements. To make
this process more robust/easier to perform going forward we are going to
encapsulate it in a oneoff service class.

Add `ChangeServiceFees` class. By default a dry run will be performed
and it will log out the changes to be made.