Extract information recording to concern
In a few of our service/lib classes that we run from the terminal its
useful to be able to record changes that have/will be made so that we
can inspect/record them.

Using `Rails.logger` for this is a pain from the terminal as we have to
wade through all the other log messages (usually a bunch from
`ActiveRecord`) or manually disable them prior to running the service
class.

Add lightweight `HasRecordableInformation` concern to be able to easily
record/playback information from a service class. Also adds a custom
matcher to clean up the testing of this functionality.