Correctly sort contracts by version
The versioning of `CallOfContract` models is not very consistent and performing
a sort operation on the version string will not always yield a correct
ordering.

Add `numerical_value` to `ContractVersion` so that we can more consistently
convert our version strings into a numerical value that we can then sort on.

Update `ChangeServiceFees` to sort on the `numerical_value`. This is less
efficient, but most lead provider/cohort combinations in production only have a
few contracts - with the current maximum being 89 - so we should be fine.

I've stopped short of pulling the logic to find the latest `CallOffContract`
into its own service to keep this set of changes manageable; we can look at
doing that when the need arises.