Display authorised for payment date in local timezone
We currently display the authorised for payment date in the same
timezone that we store it in the database (UTC). Instead, we want to
format the date and display it in the London timzone.

This is a app-wide issue, however applying a 'fix all' solution of
setting the time zone in the applicaton config is risky (it would
change it everywhere, including in our serializers). Instead, we are
going to fix this on a case-by-case basis by converting the dates into
London timezone prior to formatting them.

If we find we are going to change this in lots of other places it may
be be worth creating a date formatting service.