Optimise ect/mentor counts
Previously we were calculating the counts per-school, however as we
need to evaluate the latest induction records by lead provider as
part of this it was too slow.

Instead, the query has been updated to support multiple partnerships
and extracted into a `Participants::ActiveByPartnership` service. We
calculate this once per request and pass it down into the views.