Change delivery partner ParticipantsFilter to use induction records
As we will be querying over induction records instead of participant profiles
in the controller it makes sense to update the filter to also work over
induction records. We will be passing in the pre-calculated training record
states here as well, which is inline with how appropriate bodies work.

Includes a minor tidy up in the appropriate bodies `ParticipantsFilter` and
adds a spec file for the delivery partners `ParticipantsFilter`.