Add oneoff service to migrate declarations between statements
We occasionally need to migrate declarations between
different finance statements.

When doing this, we expect both statements to be output
statements and after migrating the from statement should be
updated to `output_fee: false`.

We only want to allow migration of NPQ declarations between
NPQ statements and only when they are of the same cohort.
We need to ensure there is a to/from statement for each
provider.