Add link to save statements as PDF
We want to let lead providers download a PDF copy of their
financial statements. To avoid the complexity of generating PDFs
locally we are instead going to leverage the browser print ->
save as PDF functionality.

When a user clicks the 'Save as PDF' link we perform some minor
page formatting via Javascript and update the page title as this
is used as the filename when saving. We also apply print-specific
CSS to hide certain elements (such as links). After the user has
finished saving the PDF we restore the page to its previous state
(although we don't remember the detail open/close state; we just
assume they were closed initially).