Account for minimum band a being set to 0
In legacy bandings, the minimum value for band A was set to `nil`, which
meant in the output calculator it defaulted to 1. Since we introduced
services to accept csvs and create contracts/bandings we set the band A
min to 0. This pushed started declarations in band A above the max value,
for instance band A max is set to 2000, it would appear in the dashboard
as 2001.

Ensure that we account for both `nil` and 0 values in calculations to
avoid wrong banding for declarations by setting both 1 otherwise the
band min value