Add make cmd for konduit to snapshot DB
Add a `make` command to more easily create a konduit to the snapshot database.

The command will remove any existing `url` in the development section of the
`database.yml` and then create a `konduit` to the snapshot DB. When its created
the `url` in the development section of the `database.yml` will be updated.

When the make command exits the `url` is removed form the `database.yml` file.