Make teacher profile selection consistent
Inside the #same_trn_user we are selecting all teacher profiles that
match a TRN but aren't specifying any explicit order, so there's no
guarantee which TeacherProfile record is found and used.

This appears to have caused some problems where providers expect a user
to have one ID but they have another, because the #deduplicate_by_trn!
method picked the 'wrong' teacher profile, so the wrong user to attach
the ParticipantIdentity to.

This change orders the TeacherProfile records by created_at, so it will
consistently link to the oldest record.