Simplify rspec output on GitHub
Currently rspec uses the `--format documentation` option, which prints out a list of the names of single test before displaying the failures. This makes it harder to find the actual failures, as there's a lot of scrolling required.

Also dropping the `--profile` option which displays the 10 slowest tests, as again this is noise, and it can always be run locally if needed.