Create analytics tables inside the main database
We've had to jump through lots of hoops to get our app, which is
unconventional in that it has two databases, to work in the regular DfE
fashion in Azure. It adds complication when dealing with the two
'setups', the one used in the main environments where PostgreSQL is a
permanent fixture ('Azure Flexible Server') and in review apps where
it's just a container.

As we no longer externally connect to the analytics database externally
but do want to stream the updates to Azure, it makes sense to move them
into the main database.