Use applications to reconcile users
If we find a user in NPQ reg has an application with ecf_id of, say, `1` and a
corresponding user in ECF has an application with id of `1` then we can infer
that these two users are the same.

Remove `email` from reconciliation process as it isn't always unique (some
users will inherit an email address, such as head@school).

It turns out this didn't actually manage to match any more user records, but
leaving it in as it may be useful later.