Add task to refresh migration DB from production
We have amigration environment that we will be using for the ECF -> NPQ
migration testing. Prior to running this in production we want to test it in a
production-like environment.

Add a GitHub action to allow us to populate the migration database from the
production database. Update `Makefile` to include command to connect to the
migration DB.