[CPDLP-2684] Amend ParticipantIdentityResolver to fetch accepted applications
We can have multiple NPQ applications to the same course with the same
provider, but only one is accepted. If there are multiple identities
linked to the user, the resolver will sometimes pick the wrong identity.
To ensure we account for multiple identities and multiple npq
applications, ensure the application is accepted.