Remove types from pull_request on deploy flow
This is an attempt to reduce the number of duplicate workflow runs that
happen on Dependabot PRs. My hunch is that the flow is kicked off at
least twice, once when the PR is opened, once when the `dependencies`
label is added and once when the `ruby` label is added, giving us 3
runs[0]

The default types are `opened`, `synchronize`, or `reopened`[1] so we're
only really removing `labeled` here.

[0] https://github.com/DFE-Digital/early-careers-framework/pull/4439/checks
[1] https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#pull_request