Fix participant query filtering by cohort
Currently, if we filter by cohort the participants query willr
eturn a user if _any_ of their induction records have a schedule
with that cohort.

Instead, we only want to return a user if their **latest**
induction record schedule has the matching cohort.

Remove `where` clause from `latest_induction_record_query` and
apply it in the `participants_for_pagination` (after joining on
the latest induction record). Remove cohort filtering from
`participants_from` method as this always receives the
`participants_for_pagination` result as its argument and so the
filtering is already applied.