Prevent duplicate NPQ applications from being created
The NPQ registration application calls an internal API in
ECF that will create an application. At the moment, the
`NPQApplication` is persisted and then the user is deduped.
Its possible for the deduping process to raise an
exception; if this happens the application will have been
persisted but the API will return a non-200 response. This
causes the NPQ registration app to retry the application
creation and results in a duplicate record.

Instead, we want to wrap this operation in a transaction so
that if any part of the application creation or dedupe
process fails the changes made will not be persisted and
the NPQ registration service can safely retry the
operation.

Adds test coverage for the `save_and_dedupe_participant` method.