Add Make command to upload files from AKS pods
We currently have a command that will download a file from a pod in a given AKS
cluster, however we have no way of uploading a file to a pod.

The current method of downloading a file will also choose a pod in the cluster
at random; ideally we want to download a file from each pod in the cluster (in
case we create the file in one pod and the command tries to download it from a
different pod in the same cluster).

Add make command to upload files and modify the command to download files to
pull from each pod in the cluster (saving the files in a subdirectory of the
pod name).