Prevent flip/flop when deduping NPQ participants
Currently, if we call `NPQP::DedupeParticipant` with an `NPQApplication` that
is associated with the oldest `TeacherProfile` it will always transfer this
user onto a more recent `TeacherProfile` (as we exclude the `from_user` from
the `to_user` lookup).

This results in users flip/flopping on `TeacherProfile` records. Instead, we
should never exclude the oldest `TeacherProfile` and only transfer if the
`NPQApplication` user is more recent.

As the `Identity::Transfer` class prevents duplicate `ParticipantIdChange`
records from being created we can omit this check from the
`NPQ::DedupeParticipant` service.