Prevent flip/flop on StoreValidationResult
When we call `StoreValidationResult` it will dedupe the participant profile by
default. If it happens to be called with the primary user (that has the oldest
`TeacherProfile`) then this can result in the user transferring to a newer,
duplicate user.

Always find the oldest `TeacherProfile` with a matching `trn` to dedupe to, and
only dedupe if that user is not the same as the one on the
`participant_profile`.