Add script to transfer away from archived users
Due to a bug where the `TeacherProfile.trn` was not being cleared on
archiving the associated user (#4524) we now have a number of
archived users with associated participant profiles.

As of #4524 we now nullify the `TeacherProfile.trn` on archiving, so
all that remains is to archived users onto the correct primary user
record.