Don't clear trn if teacher profile has declarations
Currently, we check to ensure that there are no NPQ profiles on the
`TeacherProfile` before we clear the `trn`, however we also need to check that
the associated participant profiles do not have any declarations (as
declarations should be tied back to a `TeacherProfile` with a `trn`).

Update the `ValidateDetails` and `EcfManualValidation` services to check for
declarations prior to clearing a `trn`.