Change concurrency group to head_ref then run_id
The concurrency setting in GitHub workflows will cancel a run when
another run in the same group occurs. The group was set to `github.ref`
but this appears to cancel the deployment on main before it finishes
when queues are enabled.

Instead if we use `github.head_ref || github.run_id` the cancellation
should only happen within the context of pull requests or retries

https://docs.github.com/en/actions/learn-github-actions/contexts#github-context