Don't seed the test database
We appear to be needlessly seeding the test database and then ignoring
and generating new data when we run the tests. This came to light when
we noticed tests failing when the test database had been populated via
`rails db:setup` but passing when it'd just been created with `rails
db:create`.

This change just skips seeding when the Rails environment is `test`.