Amend healthcheck to not fail if database not populated
To allow the migraiton environment database to be restored from
production there will be a certain amount of time the database is empty.
Keeping the healthcheck happy means we can keep the refresh going as
otherwise the pipeline will be broken since we rely on the migration web
container to be up and healthy.

This work is similar to NPQ thanks Ross
https://github.com/DFE-Digital/npq-registration/pull/1283