Backfill cohort on ParticipantDeclaration
We are stamping `cohort` on declaration when created, but for previously
created declarations we need to go back and backfill the `cohort`. We do this
by looking at the:

``` Declaration → statement → cohort if available Profile → latest induction
record → cohort for ECF Profile → schedule → cohort for NPQ ```

If no cohort can be found then the backfill will skip over the declaration.

We take the statement -> cohort as priority, then the profile -> schedule ->
cohort.

We do not want to touch the declaration `updated_at` when we stamp the
`cohort`.